backend/requirements.txt
fastapi
uvicorn
pymongo
motor
selenium
webdriver-manager
gspread
google-auth
geopy
phonenumbers
python-dotenv
pydantic
